+++ Makefile
@@ -1,14 +1,14 @@
-#******************************************************************************#
+# **************************************************************************** #
 #                                                                              #
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: ciglesia <ciglesia@student.42.fr>          +#+  +:+       +#+         #
+#    By: ksoto <ksoto@student.42.fr>                +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2021/05/20 17:00:07 by ciglesia          #+#    #+#              #
-#    Updated: 2021/09/14 15:55:48 by ciglesia         ###   ########.fr        #
+#    Updated: 2021/09/16 17:55:04 by ksoto            ###   ########.fr        #
 #                                                                              #
-#******************************************************************************#
+# **************************************************************************** #
 
 NAME		=	fdf
 
@@ -36,7 +36,7 @@ DIRDIS		=	$(DIRSRC)/display/
 
 SRC			=	main.c
 ANALYZE		=	analyze.c map.c
-COMPUTE		=	compute.c
+COMPUTE		=	compute.c navigate.c
 DISPLAY		=	render.c
 
 SRCS		=	$(SRC) $(ANALYZE) $(COMPUTE) $(DISPLAY)
